# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: YelaPay+ API — digital payment platform for informal transport in Angola
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [passenger, driver, admin]
        - in: query
          name: status
          schema:
            type: string
            enum: [active, suspended]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Phone number already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/wallet:
    get:
      operationId: getUserWallet
      tags: [wallets]
      summary: Get the wallet for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──────────────── Wallets ────────────────
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips
      parameters:
        - in: query
          name: routeId
          schema:
            type: integer
        - in: query
          name: driverId
          schema:
            type: integer
        - in: query
          name: vehicleId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [active, completed, cancelled]
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Start a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Trip started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Trip found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip (e.g. end it or cancel it)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/pay:
    post:
      operationId: payForTrip
      tags: [trips]
      summary: Passenger pays for a trip
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripPaymentInput"
      responses:
        "201":
          description: Payment successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripPayment"
        "400":
          description: Insufficient balance or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Trip or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{id}/payments:
    get:
      operationId: listTripPayments
      tags: [trips]
      summary: List all payments for a trip
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: List of trip payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripPayment"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──────────────── Stats ────────────────
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Users ──
    User:
      type: object
      required: [id, phone, name, role, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        status:
          type: string
          enum: [active, suspended]
        pin:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [phone, name, role]
      properties:
        phone:
          type: string
          minLength: 9
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [passenger, driver, admin]
        pin:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        status:
          type: string
          enum: [active, suspended]
        pin:
          type: string

    # ── Wallets ──
    Wallet:
      type: object
      required: [id, userId, balance, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        balance:
          type: number
          description: Balance in AOA (Angolan Kwanza)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Trip:
      type: object
      required: [id, vehicleId, routeId, driverId, status, passengerCount, totalCollected, createdAt]
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        routeId:
          type: integer
        driverId:
          type: integer
        status:
          type: string
          enum: [active, completed, cancelled]
        passengerCount:
          type: integer
        totalCollected:
          type: number
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    TripInput:
      type: object
      required: [vehicleId, routeId, driverId]
      properties:
        vehicleId:
          type: integer
        routeId:
          type: integer
        driverId:
          type: integer
    TripUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, cancelled]
    TripPayment:
      type: object
      required: [id, tripId, passengerId, amount, status, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        passengerId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        transactionId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    TripPaymentInput:
      type: object
      required: [passengerId]
      properties:
        passengerId:
          type: integer
        amount:
          type: number
          description: Override fare amount (optional, uses route fare if omitted)

    # ── Stats ──
...